+++ src/useful/libs/Useful.Audio/AudioOptions.cs
@@ -1,0 +1,10 @@
+// 'Useful Libraries' - Andy Hawkins 2025.
+
+namespace Useful.Audio;
+
+public sealed class AudioOptions
+{
+    public bool MusicOn { get; set; } = true;
+
+    public bool EffectsOn { get; set; } = true;
+}